Segments: Initial implementation
Segments are going to be parquet file that will serve as staging ground
for in-flight data. This means that ingesting replication stream with
intake won't commit data as transaction wraps, simply because it would
defeat the purpose of an OLAP/parquet workload.

The Segment holds a cache in memory of all the events to be then
compiled into a parquet format. This current implementation is highly
inefficicent just because it is storing a vec of Values
(HashMap<String, Value>) and that would mean the row to column
transformation would be completely inefficient. This is an acceptable
tradeoff because right now the goal is just to get intake to move OLTP
data into an OLAP format. Performances adjustment are going to follow
once the basic implementations are covered.